backend-teacher/requirements.txt
# Python 의존성 패키지 목록

# FastAPI 및 관련 패키지
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 데이터베이스
sqlalchemy==2.0.23
alembic==1.13.1
# PostgreSQL 드라이버 (Windows 빌드 도구 필요시 아래 주석 해제)
psycopg==3.1.18  # psycopg3 (설치 더 쉬움, Windows 친화적)

# Windows에서 psycopg2-binary 설치가 안되는 경우:
# pip install --only-binary psycopg2-binary psycopg2-binary

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 유틸리티
python-dotenv==1.0.0
httpx==0.25.2
email-validator==2.1.0

# 개발 및 테스트
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 로깅
loguru==0.7.2

# CORS
fastapi-cors==0.0.6
